requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
Cerberus==1.3.5
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.7
django-filter==23.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==20.1.0
httpie==3.2.2
idna==3.6
inflection==0.5.1
isort==5.13.0
itypes==1.2.0
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
openapi-codec==1.3.2
packaging==23.2
pep517==0.13.1
Pillow==10.1.0
pip-api==0.0.30
pipreqs==0.4.13
platformdirs==4.1.0
plette==0.4.4
pydantic==2.5.2
pydantic-core==2.14.5
pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
requirementslib==3.0.0
rich==13.7.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
yarg==0.1.9
